+++ pyocrspace/test_pictures/hehehe differentiation.png
