+++ pyocrspace/test_pictures/lmao imaginary numbers .png
