+++ pyocrspace/test_pictures/this is illegal.jpeg
